+++ lib/MIP/Cli/Mip/Analyse/Rd_rna.pm
@@ -539,6 +539,15 @@ q{Default: grch37_dbsnp_-138-.vcf, grch37_1000g_indels_-phase1-.vcf, grch37_mill
     );
 
     option(
+        q{star_fusion_aln} => (
+            cmd_tags      => [q{Analysis recipe switch}],
+            documentation => q{Detect fusion transcripts with star fusion. Alignment step},
+            is            => q{rw},
+            isa           => enum( [ 0, 1, 2 ] ),
+        )
+    );
+
+    option(
         q{star_fusion_min_junction_reads} => (
             documentation => q{STAR-Fusion: Minimum junction spanning reads},
             is            => q{rw},